Reimplement Validation of Facts DB Using Using UnexpectedValueException
Summary:
The validation endpoint merged in D57303153 caused an issue for duplicate class ValidationException. Instead of defining a new ValidationException (which fails), we now throw UnexpectedValueException

Facts database can currently lose integrity and may be invalid. In such a case we found that the same type was pointing to 2 different paths.
This diff adds a validation endpoint -> HH\Facts\validate(); that will currently throw an exception if any types lead to duplicate paths.
This can be further extended in the future for more queries that we know invalidate facts.
This endpoint also takes in an optional vector<string> of types to ignore if we want to override to not fail on certain classes we KNOW is currently a problem but not able to fix yet

Reviewed By: patzar

Differential Revision: D58226269

fbshipit-source-id: d4d79c8535e7d8a6f5fadebc07c76d4a886518b7