Add new aliases for `eden trace sl`
Summary:
`eden trace hg` is a tool to trace all the Sapling trace object fetches.
Let's change it to `eden trace sl` to keep up with the name changes from hg to sapling.
This diff adds an alias for `hg` too, then we still can use the old name of `eden trace hg`  It show up in the help too.

Reviewed By: kmancini

Differential Revision: D58165334

fbshipit-source-id: 750b2a4ec921ce133dbfc47921c2593210c3b985