cmake: Add liboqs to fizz-config
Summary:
The Fizz::Fizz cmake target public interface changes depending whether or not
liboqs was found during the configure stage. If liboqs was found during configuration, then
it is included as a dependency to Fizz. Consequently, we need to ensure that
the generated fizz-config.cmake file that our downstream consumers use also
find this dependency.

Reviewed By: abakiaydin

Differential Revision: D58302705

fbshipit-source-id: 5ad42ea7a9536675c3a0866fd56091dc66f18e33