Subclass SymbolMap and Refactor Factory Methods
Summary:
Create a helper subclass of `SymbolMap` to hold onto a `folly::Executor::KeepAlive` instead of reaching inside the map for the executor.  This now allows us to move the `drain()` call into the destructor of the subclass.  This will also handle overriding the `SymbolMap`'s executor with a manual executor when needed.

This also refactors the `make()` method to chain instead of having redundant logic.

Reviewed By: nt591

Differential Revision: D58310371

fbshipit-source-id: a3d6fb2b87cdd27fcd4c89e3bcff45a63fd99407