Introduce leaf frame to ViaIfAsyncCoroutine
Summary:
Add a leaf frame to ViaIfAsyncCoroutine and tag it as suspended when the coroutine is suspended.

ViaIfAsyncCoroutine is used to wrap a number of "leaf awaitables" that are not neither stack/executor-aware (eg. Baton, Receiver, Mutex, etc).

By adding a leaf-frame here, the debugger can identify instances of suspended stacks stuck on batons/mutexes etc.

The overhead here should be quite low (basically one extra no-inline), and is introduced only when a coroutine suspends on an I/O operation / lock (the cost of a reschedule should be much higher than the overhead introduced here).

Reviewed By: andriigrynenko

Differential Revision: D56939376

fbshipit-source-id: 3c23b65641389fedc52afb4ea2d7c3319ab0df40