Change ownership model for protocol reader/writer in cursor objects
Summary:
Eliminating the move constructions of the protocol objects more than doubles the observed CPU savings of using this feature.
We initially stored them by value to support forking read cursors, but this functionality was removed during review and there does not appear to be a need to restore it.

Reviewed By: fadimounir, Mizuchi

Differential Revision: D58386692

fbshipit-source-id: cdffc68504f597bddea6c208ccedd72e9562730e