Add a class to handle crate lookups
Summary:
This diff just wraps our `std::map<std::string, rust_crate> cratemap` into a class with a `find` member function.

Right now this doesn't provide a great deal of useful abstraction, but this new class is going to grow more complex in the rest of this stack. Starting in the next diff (D58398534) it needs to keep track of information about the top-level t_program currently being compiled, so that imports which are relative to the location of that program can be resolved. Even later in the stack, we'll be handling lookups of transitive dependencies, which fixes a longstanding deficiency in the Rust Thrift implementation.

Reviewed By: zertosh

Differential Revision: D58424898

fbshipit-source-id: 702f0a1be564d21648680dd47ed06dc8952d27ce