Uniquely identify different thrift files with the same filename
Summary:
Previously, we only dealt with Thrift files in terms of their Thrift program name, as evidenced by lookups that looked like this:

https://www.internalfb.com/code/fbsource/[ec1adb5048c62c9ea173db32f75633ef4eb34b46]/fbcode/thrift/compiler/lib/rust/util.cc?lines=106

The Thrift program name (`program->name()`) is always derived from the filename of the .thrift file. For a program parsed from `common.thrift`, the program name is `common`.

The problem is a Thrift library can have multiple direct dependencies that all contain different sources with the same program name. In this situation `program->name()` is inadequate to uniquely resolve the right Rust crate.

This diff changes the Rust Thrift implementation to use whole filepaths like `path/to/common.thrift` as the key for our crate lookups.

Reviewed By: zertosh

Differential Revision: D58398534

fbshipit-source-id: c1cabf61c6b518ef5ae49bcf7b3c0c8c16b72848