Delete usage of Eval.GetClassBadArgument from get_class and get_parent_class [1/2]
Summary:
This is a split of D58048561 in two parts after its backout. This part should be trivial and this diff can land before the others.
 ---

Removes the logging behavior from the intrinsics and removes the default null arguments. Removes the hand written type checking from get_class and uses builtin type-hint enforcement for `HH\object`.

Reviewed By: ricklavoie

Differential Revision: D58396441

fbshipit-source-id: b0bd3bf8a3b374e7eeee1e59f43e7f5f0d77ed1a