Unify LdCls* helpers
Summary:
1. Rename `LookupClsRDS` -> `LookupCls` (since they all use RDS)
2. Rename `LdClsCachedSafe` -> `LookupClsCached` since it doesn't actually trigger autoload and does same thing as `LookupCls` but with a CStr as the string argument for classname.
2. Unify implementation of `LookupClsCached` with `LdClsCached`

----------------

In the rest of the codebase, class lookup vs load differs based on whether to autoload or not.
It seems like we want 4 Ld IR instructions:
1. LdCls
2. LdClsCached
3. LookupCls
4. LookupClsCached

Load vs Lookup determines whether we autoload or not.
"Cached" means we have a constant SSA value for the String name. There may be a better name for this, since we check various caches for both `LdCls` and `LdClsCached`.

Doing this because it seems like we are not triggering the optimizations for each of these cases that we should be.

Reviewed By: ricklavoie

Differential Revision: D58037629

fbshipit-source-id: dc27d3c5f04560b480308d218631b7d73f79067e