derive Ord/Eq for structs with a map where possible
Summary: if a Thrift struct contains a `map`, we currently opt it out of deriving Ord/Eq, which makes for some confusing errors when one Thrift struct can be compared but another cannot. Rust derives `Eq` for `BTreeMap` if both the key and value derive `Eq`, so we can mimic the same logic in our Thrift codegen.

Reviewed By: shayne-fletcher, dtolnay

Differential Revision: D58441189

fbshipit-source-id: dca3becfd52f17396f7986cf07a9adb406e11cc4