move slice-catching to make_exception_ptr_with
Summary:
The `folly::exception_wrapper` constructors which take exception objects check that the static and dynamic types of the argument match to warn users when their code is likely mistaken.

Now that `folly::make_exception_ptr_with` exists and since the mentioned constructors forward to it, it is a broader place for the facility.

Reviewed By: luciang

Differential Revision: D28250136

fbshipit-source-id: 0732fb0ec5284a4bffee7ce9a8f2db7a0db2e687