Add diagnostic for using AsyncGenerator after exception or completion
Summary: Unlike `Task`, `AsyncGenerator`'s awaiter does not destroy the coroutine promise upon receiving an exception or completion. This allows the caller to resume execution after the exception or return, violating the intended control flow and arbitrary preconditions of subsequent code that should have been unreachable. As a first step towards banning that behavior this diff adds a debug-mode diagnostic for awaiting a generator that has emitted an exception or completion.

Reviewed By: ispeters

Differential Revision: D58483579

fbshipit-source-id: 8f9d5ec75b480be98288ede31cb1dd35a568cd23