Prevent simplification of LazyCls === Cls in if statements
Summary:
We allow lazy classes, class pointers, and strings to all be equal for the duration of the class pointer migration. HHBBC has a special optimization for `if ($a === $b) { }` (`Same`; `JmpZ`) that diverges from its interp for `===` (`Same`), and it was incorrectly saying a lazy class is never equal to a class pointer and vice versa.

This meant that in repo mode
```
$a = C::class;
$b = HH\classname_to_class(C::class);
var_dump($a === $b); // prints true
if ($a === $b) {
  echo "reached"; // does not print
}
```

Adds the two cases to patch this behavior.
 ---

This change is independent of {D58389447}, but both are blockers for {D58398299}.

Reviewed By: ricklavoie

Differential Revision: D58502241

fbshipit-source-id: dbfc7df5b7be0528eeb8a02e76f77ba4b61a8ddf