Deshim AutoTimer in folly
Summary:
The following rules were deshimmed:
```
//folly/experimental:auto_timer -> //folly/logging:auto_timer
```

The following headers were deshimmed:
```
folly/experimental/AutoTimer.h -> folly/logging/AutoTimer.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:deatw.folly

Reviewed By: thedavekwon

Differential Revision: D58495140

fbshipit-source-id: 19739aaf2c8c7aa69192ad112eeb98bc3a02c73b