Deshim Cleanup in folly
Summary:
The following rules were deshimmed:
```
//folly/experimental:cleanup -> //folly/futures:cleanup
```

The following headers were deshimmed:
```
folly/experimental/Cleanup.h -> folly/futures/Cleanup.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:decleanw.folly

Reviewed By: Orvid

Differential Revision: D58495337

fbshipit-source-id: 2c6eaa3a250c52bd5d38e3c09a91d471e3a5812f