Deshim PrimaryPtr in folly
Summary:
The following rules were deshimmed:
```
//folly/experimental:primary_ptr -> //folly/concurrency/memory:primary_ptr
```

The following headers were deshimmed:
```
folly/experimental/PrimaryPtr.h -> folly/concurrency/memory/PrimaryPtr.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:deppw.folly

Reviewed By: praihan

Differential Revision: D58496404

fbshipit-source-id: 99ec44f73359e85377b9101a9778aac5ed004414