Make reflection py3 unit test run
Summary:
The reflection unit tests don't run because List__ and Set__ don't exist in thrift python, so just deferring these references to testing time so that the error happens later.

Because they haven't been run so far, there was a regression with reflection functionality in auto migrate mode, because the types_reflection file wasn't getting generated, so I also fixed that.

Reviewed By: ahilger

Differential Revision: D58564247

fbshipit-source-id: e85530c3fcfe995eb3d6bcf89b94224481e3c40e