Simplify management of IOBuf SharedInfo storage
Summary:
Currently when deleting the `SharedInfo` we have to look in two places:
1) A bit flag (`kFlagFreeSharedInfo`) in `flagsAndSharedInfo_` which says if `SharedInfo` was individually allocated
2) `SharedInfo::useHeapFullStorage`, if `SharedInfo` is in the combined allocation with `HeapFullStorage`.

The two states are mutually exclusive; setting them both is an invariant violation. This state fragmentation is a result of how `IOBuf` evolved.

In fact, there are only 3 possible backing storages for `SharedInfo`: individually allocated, `HeapFullStorage`, and combined with the ext buffer. We can store this information in a single state field of `SharedInfo`, and centralize the logic that frees it.

After this is done, `kFlagFreeSharedInfo` is the last flag packed in the `SharedInfo` pointer left over, the others were gradually removed over time, so we can remove entirely the bit packing and use a plain pointer to reference the `SharedInfo`, which significantly simplifies the code.

Reviewed By: Gownta

Differential Revision: D58513030

fbshipit-source-id: 7375868f3a4b4a98826ca376ab0cebdb82cd0d40