Make JIT agree with VM for `is_a` [1/2]
Summary:
This is a split of D58073164 in two parts after its backout.

 ---

The JIT implementation of `is_a` and `is_subclass_of` does not honor the intrinsic's third argument (`$allow_string`), so it returns `true` in cases where the VM implementation would return `false`. This discrepancy didn't cause an issue because before `HH\classname_to_class`, there was no way to get a TCls value when jitting. It also didn't check for traits -- `is_a` returns false if either argument is a trait.

- Added missing traits check
- Added missing $allow_string check

Future work:

The __Native func forces the second argument to be a string, leading to unnecessary implicit conversions. Will expand it to the `classname` union and `spatch` the WWW callsites to get rid of the strings.

Background:
- jit read of allowStr deleted here https://www.internalfb.com/diff/D2634831?transaction_fbid=1844843945761068&whitespace=SHOW_ALL
- vm support for cls arguments to is_a added in D19561707
- [bug] jit cls is_a case added here D29091339

Reviewed By: jano

Differential Revision: D58389447

fbshipit-source-id: d5c3461288a4137bde7971c1f23efbc9c69cf082