revise hazptr_domain default executor not to use ThreadLocal
Summary:
A workaround for a case where it happens to be called in a deeply-nested call-stack from within thread-exit.

The default executor is `QueuedImmediateExecutor` which uses `ThreadLocal` internally. This is necessary for `QueuedImmediateExecutor`, since it is valid for there to be many concurrently-living instances of this executor type.

The `hazptr_domain` default executor does not have this requirement so, to have equivalent queued-immediate behavior, it does not have to use `ThreadLocal` internally and can use keyword `thread_local` instead.

Reviewed By: ot

Differential Revision: D58530680

fbshipit-source-id: bcec4dbc7f8a67498660b596bb41386f9dba5859