Time-based starvation protection for EventBase
Summary:
In each iteration, `EventBase` processes callbacks coming from 3 sources:

- Ready events
- Loop callbacks
- Notification queue

If callbacks from either from the latter two take too long, it can delay handling of ready events. Notification queue supports some protection in the form of `maxReadAtOnce`, but this is both insufficient and inefficient:
- Callback runtime is unbounded, so limiting the number offers no guarantee on how much they delay the iteration
- Due to this uncertainty, the default number is very low (10), which can end up unnecessarily slow down the processing by interleaving expensive calls to `epoll_wait` even if there are no ready events.

Furthermore, loop callbacks offer no such protection.

This diff introduces an opt-in mechanism to bound the time that loop callbacks and notification queue callbacks can spend in an iteration; if the deadline is hit, processing is continued in the next iteration.

The intention is to eventually replace `maxReadAtOnce`.

Reviewed By: yfeldblum

Differential Revision: D58591250

fbshipit-source-id: b7970900b1097c748d90410ac80c4870678e8768