EventBase::runAfterLoop()
Summary:
There is a pattern of using `runInLoop()` with `thisIteration=true` to schedule a callback as late as possible in the loop and use it to batch-flush operations that were deferred by other callbacks.

This pattern may not behave as intended with `loopCallbacksTimeslice` because the scheduled callback is not guaranteed anymore to run within the same iteration.

This diff adds first-class support for the pattern with `runAfterLoop()` (symmetrical to `runBeforeLoop()`), which is guaranteed to run after the `runInLoop()` callbacks (so no rescheduling hacks needed), and is not subject to the timeslice (so it must be used with care).

It nicely complements `loopCallbacksTimeslice` for the aforementioned flush pattern, since it is possible to bound how long operations are deferred (modulo individually long-running callbacks).

Differential Revision: D58610940

fbshipit-source-id: ac7ba212f18fc3750d42242daf6ba699cdb5db87