Support referencing types from transitive includes
Summary:
One of the most common unresolved sources of breakage when enabling Rust code generation on a Thrift library over the last 5 years is Rust requiring more Thrift `include` statements than what other Thrift languages require.

In C++ and Java and Python and other languages in which name resolution is based on a global namespace for types, once a type is defined within a particular namespace (C++) or at a particular package path (Java), it can be accessed at that location from anywhere that the definition gets transitively depended upon.

In Rust, name resolution is not based on a global namespace for types, which is amazing because this eliminates a big category of "dependency hell" issues that occur in those other languages. In Rust it's fine to depend on 2 different libraries (or 2 versions of the same library) which each define a type with the same name. This would be an ODR violation in C++ (UB with no diagnostic required) or classloader misadventure in Java.

Previously, Thrift-generated Rust code was only able to refer to types from direct dependencies. If the code needed to refer to a type from a transitive dependency that was not a direct dependency, the generated code would fail to compile until you filled in a direct dependency on that target.

This diff lifts this restriction by building on {D58656555} to keep track of multi-step dependency chains by which we can refer to any needed type from any transitive dependency.

Reviewed By: zertosh

Differential Revision: D58654187

fbshipit-source-id: 37a071c2b7c7fe3c4942801d167673c3c75ba07a