MutableUnionMeta: initial support, added field enum, do not raise error
Summary:
This first step towards implementing the metaclass for mutable thrift-python unions makes a notable change from the corresponding (immutable) implementation: it uses names from the reserved fbthrift space for generated identifiers, such as the enum type of all possible union fields.

This effectively addresses the issues demonstrated by the following unions in the `thrift/test/thrift-python/union_test.thrift` file:

```

# This union intentionaly uses a field name that is used by the (immutable)
# thrift-python union class implementation to expose an enumeration with all
# possible fields. See union_test.py for  the resulting behavior.
union TestUnionAmbiguousTypeFieldName {
  1: i32 Type;
}

# This union intentionaly uses field names that are used internally by the
# (immutable) thrift-python union class implementation. See union_test.py for
# the resulting behavior.
union TestUnionAmbiguousValueFieldName {
  1: i32 type;
  2: i32 value;
}
# NOTE: Error on import (due to fake "EMPTY" enum value):
# TypeError: Attempted to reuse key: 'EMPTY'
# union TestUnionWithEmptyFieldName {
#   1: i32 EMPTY;
# }
```

Obviously this comes at the cost of API difference between the immutable and mutable API, which is not desirable. It is therefore still an open question whether this approach will be the final one, but given that mutable API support is still very much experimental, I suggest proceeding with this for now and discussing any changes/migrations offline.

Reviewed By: yoney

Differential Revision: D58210725

fbshipit-source-id: 9251c7cb35fa9be5a49be546f0b9178687e74294