Fix AtomicNotificationQueue CONSUME_STOP behavior with interleaved pushes
Summary:
`CONSUME_STOP` (introduced in D58591249) is incorrect because the atomic queue is moved eagerly based on the assumption that `maxReadAtOnce_` (or all) tasks will be processed, which is not true if one task stops, so the next queue may be leaked.

The eager move (instead of refilling `queue_` once empty) is there to guarantee that tasks put during the execution of `drive()` are not processed, so `drive()` returns in bounded time even if `maxReadAtOnce_` is `0`. I'm not sure if this behavior is actually desirable (and I plan to revisit it in the future), but for now we should not break it.

We can keep it by marking the last task in the atomic queue when we enter `drive()`, so we know when to stop. This also fixes the unexpected behavior described in the comment, where `drive()` may terminate early due to discarded tasks.

Reviewed By: Gownta

Differential Revision: D58647794

fbshipit-source-id: afb9d1a4b76034508d7a3ddb72442301b8e805c8