Fix TSAN errors in ManualTimekeeper
Summary:
TSAN complains about `ManualTimekeeper` because of how `now_` is updated when it's bumped in coroutines, because technically those may run on several threads, even if everything's perfectly synchronized.

To fix this, I'm making `now_` atomic. It's not a very elegantly coded solution, but `ManualTimekeeper` is only ever used in tests so we shouldn't worry about having a CAS loop in there.

Reviewed By: yfeldblum

Differential Revision: D58640261

fbshipit-source-id: e7863240bee2cff432d54888fd01c9480cc5da04