Empty tuple values
Summary:
To properly support arity polymorphism we need to allow empty tuple construction i.e. `tuple()`. This is currently outlawed in the naming phase, but is supported by HHVM, so we just need to remove the check.

Note that we still do not support parsing of the empty tuple *type* i.e. `()`.

Differential Revision: D57908217

fbshipit-source-id: 10fa0d16da8703a6bd6746f879fae91e95511308