type_predicate - Add IsTupleOf
Summary:
This diff adds a new predicate variant `IsTupleOf`.

There's a small refactor to move the partition_f into the split_ty function such that it can make use of the predicate and functions more like a helper function.

The logic for splitting a tuple type by a tuple predicate is mostly handled in the Partition module (D57976854)

Partitioning the negation of a tuple predicate by another tuple predicate is done conservatively using a simple span in this diff. Stacked diffs attempt to handle this case better.

Reviewed By: dlreeves

Differential Revision: D57623567

fbshipit-source-id: 12fb7e52f11437ece77b0d07e13560185cef06ad