Fix red squiggles, completion, and go-to-def for XHP elements in notebooks
Summary:
## What

Bug: XHP elements used to always have red squiggles in notebooks.
The fix is to properly handle namespaces and to not elaborate top-level statements twice.

**Before:**

 https://pxl.cl/535n2

**After**:

 https://pxl.cl/535LF

## Background

We wrap top-level statements in a function before type-checking, here's why: https://www.internalfb.com/intern/wiki/Hack_Developer_Guide/Hack_Notebooks/Hack_Notebooks_Top-level_Statements/

## How

Before this change, we elaborated top-level statements twice: once before wrapping in a function and once after.

After this change, we only elaborate once. And we take more care to propagate parser options to the namespace environment. The `ns_disable_xhp_element_mangling` setting turned out to be important.

As a result of this change, we also move the logic for propagating `ns_uses`. The logic used to be in `Naming_elaborate_namespaces_endo` but is now in `Naming` since we deleted the duplicate elaboration pass and our propagate-parser-options-to-namespace logic is in one place.

## Follow-on work

More code is deleted later in the stack

Differential Revision: D58410413

fbshipit-source-id: 0e221d5b7d93d54a7874f1317334dee1a817981f