Extend JIT for `is_a` to cover strings and lazy classes [2/2]
Summary:
This is a split of D58073164 in two parts after its backout.
 ---
The NativeImpls for `is_a` and `is_subclass_of` support strings and lazy classes for the $object parameter. For them, through the `get_cls` helper, we call `Class::load` on the argument, and return false if it returns nullptr. The IR instructions for class loading do not support returning null, so we add a new fallback variant `Silent`.

- Add new dest kind DLdCls that is either `TCls` or `TCls|TNullptr` based on the fallback extra-data
- Use ldCls helper with new `Silent` fallback for irgen of is_a
- Reordered + simplified the conditions a bit for NativeImpl
- Update docs to spell out the trait behavior as it is not well documented https://www.php.net/manual/en/function.is-a.php#127930

Reviewed By: jano

Differential Revision: D58392241

fbshipit-source-id: f0ace6b5d1fab445a8e0c4424f7b82f4a448c9c5