Introduce Tlabel Locl Type and Thas_const Constraint Type
Summary:
# Context
When Enum Class Labels were implemented, they were restricted to being used in only function call position. This restriction was in place because it utilized a specialized form of inference that is implemented within `Typing.EnumClassLabelOps`. We want to lift this restriction and utilize more standard inference techniques for two reasons:

1. There is a desire to utilize enum class labels for other use cases besides passing immediately to a function. For instance, having a vec of labels that are iterated over and used to invoke a function or to use labels as the type of an XHP attribute.

2. The type constant in enum class feature is currently unsound. This was attempted to be fixed in D54439754, but that approach is incomplete. For a more complete solution we need more context than is available using the current `Typing.EnumClassLabelOps.expand` function

There is also a desire to explore usage of labels for other purposes besides mapping to enum class labels.

# Solution

This stack of diffs deletes the ad-hoc `Typing.EnumClassLabelOps` based inference for enum class labels. Instead it introduces a new locl type `Tlabel` and a new constraint type `Thas_const`.

## Definition
The new locl type is defined as `Tlabel of string` where `string` is the name of the label. I.e. for the expression `#Foo`, it would have the type `Tlabel "Foo"`.

The new constraint type is defined as `Thas_const of { name: string; ty: locl_ty }` where `name` is the name of the class constant and `ty` is the type ascribed to the class constant.

## Subtyping
The `Tlabel` type has a subtype relationship with the `HH\EnumClass\Label` type. Specifically

```
E <: HasConst(A, HH\MemberOf<E, T>)
---------------------------------- :: Label-Sub-EnumClassLabel
#A <: HH\EnumClass\Label<E, T>
```

`Label-Sub-EnumClassLabel` states a label `#A` is a subtype of an `EnumClass\Label<E, T>` if `E` has a class constant named `A` whose declared type is a subtype of `MemberOf<E, T>`.

The other subtyping rule added is:
```
LookUpConstType(C, Id) <: T
----------------------- :: Class-Sub-HasConst
C <: HasConst(Id, T)
```
`Class-Sub-HasConst` states assuming `C` is a class type, it is a subtype of `HasConst(Id, T)` if the declaration of `C` contains a class constant `Id` whose declared type is a subtype of `T`. Though not stated in this rule, `LookUpConstType` can reference the `this_ty` that is bound in the environment so we can properly instantiate the `this` type when localizing the decl type associated with `C::Id`. This is what allows us to address the incompleteness introduced by D54439754

## Typing
With the above typing rules we can simply give the expression `#A` the type `Tlabel "A"` and rely on the above subtyping rules to give us the behavior we desire. This allows us to give a type to an anonymous label expression at any position, meaning we can lift the restriction on them appearing only as function arguments

# Alternatives Considered
Adding a new locl type isn't a small change and I considered alternatives to avoid having to do this. I started with a single `Thas_label` constraint type that is similar to the current `Thas_const` type. In typing I would then generate a type `EnumClass\Label<#1, #2>` with the added constraint of `#1  <: Thas_label(Id, #2)`. While it worked mostly well, it ran into issues due to the second type parameter of `EnumClass\Label` being invariant. This meant we could never early solve that type variable and led to incompleteness.

I tried a few other changes attempts, but none had the desired affect so I settled on adding the `Tlabel` type as this diff does.

# This Diff
1. In this diff I add the `Tlabel` and `Thas_const` types
2. In D55829619 I implement the subtyping and typing changes described above

Reviewed By: viratyosin

Differential Revision: D55829620

fbshipit-source-id: 43c7c30280d83f669025fe0d34248c67d54c94d0