Explicitly enumerate valid nameof targets in lowerer
Summary: Expecting the result to be an `Aast.Id` is insufficient because it does not see parens e.g. `nameof (Foo)`. This is tricky in the parser because xhp names, self, static, parent, and (qualified) names are heterogenous in their parsing. Let's just be explicit about what Syntax nodes are supported.

Reviewed By: viratyosin

Differential Revision: D58843483

fbshipit-source-id: 383ed114a8f1dd7bff94879fc05e96d0efc05a3f