Multifile uses Relative_path.create_detect_prefix instead of assuming Dummy as the prefix
Summary: There are test cases using multifile that care about the relative path as opposed to the absolute path. This makes it necessary to have the prefix and the suffix portions known within a full path. Change the creation sites in multifile to retain the accurate prefix (root, hhi, tmp) if one is set.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D58838823

fbshipit-source-id: 525fa4b01d7530927f8bb279abdba4582cf8df63