Deshim TestUtil in hphp
Summary:
The following rules were deshimmed:
```
//folly/experimental:test_util -> //folly/testing:test_util
```

The following headers were deshimmed:
```
folly/experimental/TestUtil.h -> folly/testing/TestUtil.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:detuv.hphp

Reviewed By: dneiter

Differential Revision: D58833793

fbshipit-source-id: 33b600fb683658389323c97fc1d065ebb5019428