Move ExecutionObserver to folly/executors
Summary:
The following targets were moved to folly/executors:
```
//folly/experimental:execution_observer
```

`arc f` was applied

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:ad_meeo.folly

Reviewed By: yfeldblum, Orvid

Differential Revision: D58028038

fbshipit-source-id: 0fd1dfb349421d61e9f49eb546b4c4cc0d0c0c19