Deshim TestUtil in mcrouter
Summary:
The following rules were deshimmed:
```
//folly/experimental:test_util -> //folly/testing:test_util
```

The following headers were deshimmed:
```
folly/experimental/TestUtil.h -> folly/testing/TestUtil.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:detuv.mcrouter

Reviewed By: praihan

Differential Revision: D58833810

fbshipit-source-id: 4f150bbd1ce273bda2b340d45ea9fe2f9a4ceaca