Generalize and optimize MeteredExecutor
Summary:
`MeteredExecutor`, as implemented, can introduce a noticeable concurrency bottleneck due to its serializing of the enqueue on the wrapped executor. Any work done in the worker before scheduling the next worker is in the critical path, so we need to make it as fast as possible.

This currently includes dequeueing from the notification queue (note that `scheduleCallback()` happens after that). To reduce latency, instead, do all accounting operations on a single state word so we can schedule the next worker before dequeueing the task.

Furthermore, the existing implementation ends up scheduling workers even when the queue is empty, wasting work on the wrapped executor: this is because of the condition `queue_.drive(consumer) || !queue_.arm()`, which is true whenever we processed a task, even if there is no other task in the queue. This behavior was somehow even sanctioned in the `CostOfMeteredExecutors` test, despite being undesirable. The new implementation correctly only schedules one work per submitted task by doing atomic accounting in the state word.

Finally, reintroduce `maxReadAtOnce`, removed in D33112001, but renamed `maxInQueue` for more clarity. The new implementation makes it easier to support it, and benchmarks show that increasing it from the default 1 can significantly reduce the bottleneck on high-parallelism executors.

Reviewed By: Gownta

Differential Revision: D58999318

fbshipit-source-id: 1e78855f603805ddc988eb55906c4755ca9ee376