Make Redundant Generics in EnumClass\Label an Error
Summary:
# Context
In D58215741 I changed the variance of EnumClass\Label. As a result of this change there could exist generics which are now redundant because they are under constrained. I want to make this an error because when using labels, it is very likely this is a programming bug.

Currently we only check for redundant generics that appear in a positive position of a function (i.e. output position). However this isn't enough to capture this potential dangerous pattern

```
function set<T>(HH\EnumClass\Label<E, T> $label, T $value): void {
  ...
}
```
Where the intention is to ensure that `$value` has the same type as the value referenced by enum class label `$label`. This requires reporting redundant generics in negative positions.

# Mitigation

The `redundant_generics_check.ml` TAST check is duplicated from the `linter_redundant_generics.ml` linter but only checks for generics in the positive position. I copied over the code from the linter that checks for generics in negative positions, but only perform this check if the generic appears within a `HH\EnumClass\Label<_, _>` type.

Reviewed By: andrewjkennedy

Differential Revision: D58960960

fbshipit-source-id: b71e5e9454a4cbbb6e443e775bc6b95f703cffaf