document thread safety of PendingChanges
Summary:
There was some conern on D58319081 about the thread safety around interacting
with PendingChanges. PendingChanges is a non thread safe class that should
only be used accross threads when wrapped in a PendingCollection object which
adds a lock around PendingChanges.

While landing D58319081 I did a review of all the usages of PendingChanges
and confirmed that we only directlt use PendingChanges for local objects that
are not shared accross threads, and use PendingCollection for the objects that
are shared accross threads.

Documenting these assumotions/assertions, so next time folks don't have to go
digging to confirm changes are safe/how to use pendingChanges.

Reviewed By: chadaustin

Differential Revision: D58874824

fbshipit-source-id: 0483732b3ca920c9b48538ec352592ba37e87146