Deshim TestUtil in fizz
Summary:
The following rules were deshimmed:
```
//folly/experimental:test_util -> //folly/testing:test_util
```

The following headers were deshimmed:
```
folly/experimental/TestUtil.h -> folly/testing/TestUtil.h
```

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:detuv.fizz

Reviewed By: knekritz

Differential Revision: D58833787

fbshipit-source-id: 5c9d5b227d7d795d1632369a9971d19b2dfa2a8c