Move observer/detail to folly/observer/detail
Summary:
The following targets were moved to folly/observer/detail:
```
//folly/experimental/observer/detail:graph_cycle_detector
//folly/experimental/observer/detail:observer_manager
```

`arc f` was applied

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:meod.folly

Reviewed By: Orvid

Differential Revision: D59012137

fbshipit-source-id: 34c304c4e571bc8c3e43a045540525d32c151cb2