Move huge_page_util to folly/io/tool
Summary:
The following targets were moved to folly/io/tool:
```
//folly/experimental/io/tool:huge_page_util
```

`arc f` was applied

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:meithpu.folly

Reviewed By: Orvid

Differential Revision: D59012689

fbshipit-source-id: 569e07c637889296b38478e2ef6af11fb5eac5ca