Ensure block hint is applied correctly
Summary:
While looking at TC prints of sandbox mode I noticed most of the code we emit ends up in cold. It turns out the culprit was [an unlikely hint](https://www.internalfb.com/code/fbsource/[b2ecbf5ca7d8df8511ff57d3311b4882a812cac3]/fbcode/hphp/runtime/vm/jit/irgen.cpp?lines=214) we generate when checking if code coverage is enabled. The call to `hint()` occurs after we generate a block ending instruction but before any instructions are generated in the `->next()` block so it ends up being applied to the wrong block.

With this change we'll try to start the next block before setting the hint. Most users of `hint()` are in `->taken()` blocks and so don't have this problem, but the cold path for check coverage is actually the `->next()` block because of the way `CheckRDSInitialized` is used.

Reviewed By: ricklavoie

Differential Revision: D58614053

fbshipit-source-id: 2e4bbc96bd0530cb70926d2d0e2e9a71ad1f6b85