TCPrint for sandbox mode
Summary:
When TC dumps are enabled in non-repo mode we need to dump a repo file that can be used when reloading the dump, so change `runtime-compiler.cpp` to store serialized UEs to be written to disk during a dump when the flags are enabled (and repo mode is disabled).

Also add a new script `tc-local-driver.sh` for local collection and allow `TCP_COMMAND` to be used to collect until a command completes (this is useful for benchmarking scripts).

Reviewed By: voorka

Differential Revision: D58301537

fbshipit-source-id: 2b2a89ec7ccb9e2cb04523857511832c75a3392c