Add Aliasing to Case Type Fuzz Testing
Summary:
To further test the runtime implementation for case types I added an extra layer of indirection where I assigned a case type to a type alias and then reference the case type via the alias. This uncovered another assertion failure that can occur when one of the variants is an enum.

See P1444547077 for an example trace

The broken test instances are checked in under `hphp/test/slow/case-types/generated-fuzz-test-broken-alias-*.php.disable`

The assertion occurs when the test is run using `hphp/test/run -r`

Reviewed By: vassilmladenov

Differential Revision: D59027160

fbshipit-source-id: 0b5f1e1f0ab631306a808acc512e20c70f8e149f