ThreadLocal: narrow accessAllThreads() lock scope and eliminate isLinked.
Summary:
accessAllThreads() is used to walk a specific TL id within a TL tag.
However, it locks out changing any TL object at the tag level. This happens
because the reset() calls try to acquire the accessAllThreadsLock_ that is held
by the accessAllThreads() created Accessor object. Since we now have per id
tracking set (with its own lock), we can use that to co-orindate between the
Accessor and any thread trying to alter its TL object. This will be at the id
level and lower interference from Accessors in cases where the same tag is used
by a lot of ThreadLocals -- this is common case for stats objects.

Also removed the recently added isLinked field which was used to avoid overhead
of updating the tracking set on each reset. The tracking set has to be locked
in read mode on each reset (for per id coordination with accessAllThreads), and
looking it up for presence of a thread entry is cheap, we don't need a separate
isLinked field that must be kept in sync. Removing it also prepares to free up
16 more bytes from ElementWrapper.

Reviewed By: yfeldblum

Differential Revision: D58906038

fbshipit-source-id: d25afaec4249456d258bb921c21404fc489df273