Change lock order expecation between StaticMeta.lock_ and ThreadEntrySet lock.
Summary:
The dispose() of a thread local copy of an object is called when a
thread does a reset() on it. It can invoke arbitrary code including calling
destroy() or reset() on other ThreadLocals. Such a call will need to acquire
the StaticMeta's lock_. To enforce that lock order, we need to fix paths that
today acquire a ThreadEntrySet lock after acquiring StaticMeta's lock_.

Reviewed By: yfeldblum

Differential Revision: D59076052

fbshipit-source-id: 0109d66b3a5f74d7460b8a2d27103f2e919677e9