further shrink ElementWrapper for ThreadLocal
Summary:
Use the low bit of the deleter as the tag between pointer-to-function and pointer-to-heap-allocated-callable, and the next bit as the tag for is-linked. This allows removal of the extra bool fields, reducing the size of the `ElementWrapper` structure from 24 bytes to 16.

Each thread has a vector of `ElementWrapper` of size proportional to the largest id, so the aggregate size of the vectors is `O( num-objects × num-threads )`. This reduces the constant factor by a further third.

This follows {D55115468} by ppanchalia, which reduced the size of the `ElementWrapper` structure from 56 bytes to 24.

Differential Revision: D59028606

fbshipit-source-id: 9c6e651f34956a585e0fa488963f2aad45d61db3