Modify the Thrift compiler to check for reserved identifier names in the Thrift IDL
Summary:
In addition to reserved words, Thrift reserves all words that begin with the case-insensitive string `fbthrift` preceded by zero or more underscores and should not be used as identifiers. The use of such names as identifiers requires explicit annotation of the usage with `thrift.AllowReservedIdentifierName`, and may result in undefined behavior.

Implementation notes:
The Thrift compiler will return an error when it encounters any identifier name that violates the above restriction unless the identifier is annotated with `thrift.AllowReservedIdentifierName`.

Reviewed By: iahs

Differential Revision: D59029750

fbshipit-source-id: 242dc7d71e1915bc7ffcffb6b156e49e7a6cbee3