Include program checksum in hash calculations
Summary:
To catch the merging of schemas from different snapshots before they
result in data corruption, we include the checksum of the program contents in
the hash calculation for that program and all of its definitions.
This ensures the hashes are still stable across runs within a snapshot, but
will cause resolution to fail if multiple snapshots are combined.

Reviewed By: praihan

Differential Revision: D59247590

fbshipit-source-id: a0971c9c86faf108f31c880094b7113f412f678c