map memokeys to an int instead of obj address
Summary:
So far we use the address of the object as the memo-key. This works because we intern the objects and for a given string memo key, there will be only a single object representing its IC.

The planned changes for MCP in the followup diffs will cause that assumption no longer to be true. We could be duplicating a context multiple times since we want to keep the memo-sensitive and memo-agnostic contexts alive and creating one will require duplicating the other

This diff maps the string keys to an int using the sidemap. The key 0 is reserved for all memo-agnostic ICs by means of an constexpr. Memo-sensitive keys start from 1.

Reviewed By: jano

Differential Revision: D58119734

fbshipit-source-id: a7e3a10f925d1e13aa9b24e2b5efa979499322e0