EnumClass\Label Should Not Allow References to Type Constants
Summary:
# Context
For every type constant `T` we create a constant definition with the same name as the type constant with type `TypeStructure<this::T>`. When type constants were added to enum classes, the label code did not filter out these constants. The new label inference repeated this mistake. This diff is meant to fix it.

# Mitigation
The fix is to filter out type constants within Typing_subtype.Has_const

While doing this I realized there was a bug introduced by D55829619 where if no quick fixes were produced we wouldn't report an error. I fixed this and added a test case

Reviewed By: francesco-zappa-nardelli

Differential Revision: D59340204

fbshipit-source-id: b64d868a62441e3a767795ad2b8ce83312a810fb