Suppress `supportdyn` in rendered reasons
Summary:
`supportdyn<T>` is implemented as a newtype but is not fully expanded to it's defintiion (`T & dynamic`) during constraint simplification. Coupled with it being implicitly applied, the associated extended reasons end up looking a little odd.

This diff post-processes `path`s to remove witnesses related to `supportdyn` and to correctly apply projections to the inner types.

Reviewed By: madgen

Differential Revision: D59063067

fbshipit-source-id: 62ede464da711e206d2bd6bd09fa92d55d0abe23