Render simplified path for debug output
Summary:
When constructing reasons we build up a tree structure and avoid performing reversals of paths. Before explaining the reason, we always normalize it (i.e. perform reversals) then flatten it into a alternating sequence of nodes (simple witnesses) and edges (flow direction & an optional projection)

This output is generally easier to read so we switch to using it when running under the debug flat

Reviewed By: andrewjkennedy

Differential Revision: D59374827

fbshipit-source-id: 441c77b22c19fd69431fae52160bae4dc24b7601