Use namespace instead of struct to hold program constants
Summary:
There are several pairs of files with the same filename and cpp2
namespace in the wild, which in the current codegen restricts them to at most
one of those files having constants as the containing struct would otherwise be
redefined. Instead, using a namespace causes the consts from the two files to
be merged and only has a redefinition error if the names of the consts also
match.

Reviewed By: yoney

Differential Revision: D59248697

fbshipit-source-id: cf5560d0f465d66720785efa43511e072d71e761