disable Union __dict__
Summary:
Setting `__slots__` on Union may make it faster to initialize and reduce its memory footprint.

This shouldn't meaningfully affect performance because the changes are all in the metaclass (import time).

At diff time the difference is `__slots__` vs `__dict__`, which is mainly a memory usage difference, with `__slots__` being more efficient.

Reviewed By: yoney, Filip-F

Differential Revision: D59599431

fbshipit-source-id: 5c356a67c61879c4584d16235488370cb3dd2b2a