Move coding_test_utils to folly/compression/test
Summary:
The following targets were moved to folly/compression/test:
```
//folly/experimental/test:coding_test_utils
```

`arc f` was applied

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:metctu2.folly

Differential Revision: D59596566

fbshipit-source-id: 344201d586a931d2579ba8e814c25a1f2af8edbe