Move cli_example to folly/cli/test
Summary:
The following targets were moved to folly/cli/test:
```
//folly/experimental/test:nested_command_line_app_example
```

`arc f` was applied

This is a codemod. It was automatically generated and will be landed once it is approved and tests are passing in sandcastle.
You have been added as a reviewer by Sentinel or Butterfly.

p:metclie.folly

Differential Revision: D59606045

fbshipit-source-id: b676d135bd7dac14461bec44188d680fa26e8c7f