Load classId out of RDS for EqClassId
Summary:
We'd like to store the classId for a given class name in RDS, so that when we go to do the runtime checks about whether we can trust a method/constant/property, we can just do an RDS check rather than having to perform a `mov` instruction on a `Class*`

RDS has a few ways of binding.
Either you `bindLink()` and the link declaration is unique.
Or you regular `bind(sym)` and the symbol is unique.

In this case, we want the RDS handle to be shared among all links with the same symbol (ie all classes with same name should have the same handle).

----

A follow-up to this could be switching from a ClassId to a classVersion. The benefit here would be that we could store it in a uint8_t, at the expense of maintaining some sort of global map with atomic counters per Classname to figure out what the next version to use is.
I'm not certain this would really save that much, or be worth the extra global table to keep track of these atomic counters.

Reviewed By: jano

Differential Revision: D59502140

fbshipit-source-id: 3b6cd4afc656f60b99b1579e624f68d0e66a2901