Fix parser precedence bug
Summary:
When parsing expressions with binary operators, special care has to be taken to tell the difference between `<` used as a comparison binop, and `f<t1>`, `f<>` syntax for generic type arguments, and function pointer creation. Essentially, when encountering `<` we try to parse the function call/pointer expression and only treat it as comparison when the former fails.

In a case like
```10 - f<>(1) - 1```
we'd detect the situation and correctly parse the call to `f`. But then we'd continue to parse through the chain of operators as though the next one (the second `-` in the example) was higher precedence or right associative, without actually looking at it.

Here, we recursively call the parsing function having just extended the right term from `f` to `f<>(1)`, and queued up the second `-`, which then causes the parser to take the left associative path. So we compute 8 instead of 10.

The buggy path was introduced in D24365459 to fix another bug where `<` in function calls made the precedence wrong.

Reviewed By: madgen

Differential Revision: D59685896

fbshipit-source-id: 20458421f7ddb1656ab96c4956fd87d629f342cd