Move schema mutator after standard library load
Summary:
We currently run the mutator before the standard library is loaded, which means the constant injection fails silently for files that don't include schema.thrift. Moving this mutator after the load causes it to succeed for all files.
To prevent these silent failures in the future, we gate the const injection behind a flag and cause it to fail the build when it doesn't succeed.

Reviewed By: praihan

Differential Revision: D59656228

fbshipit-source-id: 15ac7fbabcfdab91eb2824720058c0e6264a5e57