Support <<__Overlapping>> attribute on methods
Summary:
The existing `<<__NonDisjoint>>` generic parameter attribute only supports function-level and method-level generic parameters. Disjointness with enclosing class-level generics has to be implemented via additional method-level generics and `where Tv2=Tv` clauses, which means interfering with the API unnecessarily.

Here we implement an alternative approach using a method attribute `<<__Overlapping('T1', ..., 'Tn')>>` that specifies a set of generic parameters, from both the enclosing class and the method, whose arguments are expected to be disjoint types.

For now, the attribute is supported only for methods (instance and static). To support it for top-level functions requires a little more work on decls.

Reviewed By: madgen

Differential Revision: D59754309

fbshipit-source-id: f02f8d85859ec182caa18a0b5adac942931988f3