Add new shim types for floating point version of `folly::toAppend`
Summary:
These new shim types mimc the types of double-conversion. This diff simply adds these new types but does not yet use them. Usage will come in a separate diff.

This diff is the first step of a multistep removal process of double-conversion. This first step adds new shim types. Later steps will update callers to replace the double-conversion enum types with these new types. Eventually double-conversion will be removed entirely and replaced with an implementation that uses `std::to_chars` and `std::from_chars`.

These new types decouple folly from double-conversion.

double-conversion is removed because 1/ it increases android .apk binary sizes by ~100KB due to the inclusion of its libdouble-conversion.so (T194496993) and 2/ there is equivalent functionality in the standard library that eliminates the need for this library.

I keep the interface the same to make the transition to a new implementation as simple as possible.

Reviewed By: Gownta

Differential Revision: D59602063

fbshipit-source-id: 13539783f862a6098ad6353ac0e8dde64a170122