Use folly::DtoaMode for folly::toAppend
Summary:
Use `folly::DtoaMode` for calls to `folly::toAppend`. The behavior is expected to be the same.

This is a step in the multistep process to remove `double_conversion`. This step switches `toAppend` callers to use the shim type.

double-conversion is removed because 1/ it increases android .apk binary sizes by ~100KB due to the inclusion of its libdouble-conversion.so (T194496993) and 2/ there is equivalent functionality in the standard library that eliminates the need for this library.

Reviewed By: Gownta

Differential Revision: D59770924

fbshipit-source-id: 7fb7af2164b8cb0eaf41d8de64ec9060daf981b6