Push reversals through 'Typing_reason.to_string'
Summary: The `Rrev` constructor is used to delay reversing a path defined by an extended reason.  When reporting the position based on a reason, we want the leftmost witness. Since the leftmost witness of `Rrev t` is the rightmost witness in `t` so we  to change directions. This diff applies the same logic for the `to_string` function which gives an explanation of the relevant witness for legacy error reporting

Reviewed By: CatherineGasnier

Differential Revision: D59676304

fbshipit-source-id: 8845589e35ad65e384e74981d0693233fdadfa82