Remove routing hint (tech debt)
Summary:
This was an optimization we introduced for Memcache back in 2021. The idea was to pass the host returned by `getConsistentHashHost()` back to the routing layer (SR) as a "routing hint". SR could then avoid performing host selection again.

Unfortunately this optimization did not play well with failovers, and we were forced to disable it almost immediately. It was quickly forgotten and became tech debt.

I don't think there is any value in trying to rescue this optimization. The cost (in terms of engineering hours and code complexity) is probably not worth any potential gains in performance, if any. Today SR uses a dedicated uCache route handle tree, which includes a streamlined connection pool implementation, so any gains from a routing hint would be minimal or nil.

Reviewed By: stuclar

Differential Revision:
D56376007

Privacy Context Container: L1225223

fbshipit-source-id: 45ed61a4bf255c14368d44076aa71f01d8d6d9ed