ClientInterceptorBase
Summary:
This diff does two things:
1. Introduce `ClientInterceptorBase` (similar to `ServiceInterceptorBase`).
2. Manage `ClientInterceptorBase` state + storage in `ContextStack` (via which these will be invoked).

At present, this list of `ClientInterceptorBase` passed to `ContextStack` will be empty and thus this diff has no observable side-effects.

`ContextStack::create*` has an optimization where it returns `nullptr` if the set of `TProcessorEventHandler`s is empty. We also changed this optimization to take into account the set of provided `ClientInterceptorBase`.

Reviewed By: thedavekwon

Differential Revision: D59661921

fbshipit-source-id: f525f0caa43c402cc0cc67e92b85c20189c81023