Fix crash related to resolving type aliases that include an enum.
Summary:
Attempt to fix test case `fbcode/hphp/test/slow/case-types/generated-fuzz-test-broken-alias-1.php`
It looks like what was happening was we we had some code like:
```
<?hh
  /**
   * THIS FILE IS generated; DO NOT EDIT IT
   * To regenerate this file, run
   *
   *   buck run //hphp/hack/test:gen_case_type_tests
   **/

  <<file: __EnableUnstableFeatures('case_types')>>

  enum MyEnum : string {
    A = 'A';
    B = 'B';
  }

  case type CT0 = MyEnum|dict<int>;
  type AliasCT0 = CT0;

class CheckAliasCT0 {
   private AliasCT0 $field;
}

<<__EntryPoint>>
  function main(): void {
}
```

When we went to call `update_type_constraint()` on `AliasCTO`, we iterated over all tcs in CTO, and called `resolve_one` on each of them. We passed in the firstEnum, `MyEnum` to each of them as the `clsName` (still not sure why we do this). This created a `TypeConstraint` for each type in the `CTO` case type with a `clsName` `MyEnum` with the original type (string, and dict<int>). This triggered an assert in debug mode because we failed when we ran `enumSupportsAnnot(Dict)`, since `dict<int>` is not allowed to be an `enumBaseTy`.

This may not be the right fix. See comments below.

# Fix rationale:
It seems like if we just need `firstEnum` for missing types later on, these are set earlier on the `TypeMapping`, and whether we pass them to `resolveType()` or not, they still remain on the `TypeMapping` to be used later.

Reviewed By: ricklavoie

Differential Revision: D59606610

fbshipit-source-id: 6094ad40bcde19d1b08f95c16fc21ebc9efcc294