`fizz::DefaultFactory` aliases to OpenSSLFactory
Summary:
Prior to this diff:
* `fizz::Factory` defined only the interface.
* `fizz::DefaultFactory` defines certain components that should normally be
  agnostic to specific backends
* Actual concrete factory implementations are supposed to extend `fizz::DefaultFactory`

The additional `fizz::DefaultFactory` step is unnecessary. This diff effectively
reverts most of the changes in D51481217.

`fizz::DefaultFactory` now is a type alias to a _specific concrete factory_ that
was configured at compile time. Other internal parts of Fizz never refer to
a specific factory, but rather use `fizz::DefaultFactory`.

Differential Revision: D59615209

fbshipit-source-id: 9fb11863eeb6defcc79b501e6f054b91c1fd81ac