simplify cache local/lifetime coordination in SingletonRelaxedCounter
Summary:
Rather than giving up on constraining the cache thread_local local and cache thread_local lifetime to be in the same DSO, now constrain it and assert on the constraint. This allows the cache thread_local local and the cache thread_local lifetime to be 1:1 with each other, rather than M:N with each other without the constraint.

Note that there is now no cross-DSO merging of these thread_local variables - they are now explicitly per-DSO.

Differential Revision: D59144274

fbshipit-source-id: 73b3a426059321bb9abfcb92fdbdd3547656adc8