use THP when 1G pages are unavailable on aarch64
Summary:
On aarch64, THP size can be something like 512MB (or maybe 32MB depending on configuration). So we use THP in place of configured 1GB pages.

This is kind of a hack in that we are not introducing options for the correct page size, and instead use THP to back what was supposed to be 1G pages. However, in practice, the THP size on aarch64 is either 512MB or 32MB, and using such pages when 1G pages are absent makes sense, and it saves us more complicated options.

Reviewed By: ricklavoie

Differential Revision: D60003426

fbshipit-source-id: 70e4139dccb74e5c5fe7eacad7bb4184de69c9a5