Remove instantiated_targs flag from function types
Summary:
The `instantiated_targs` flag is redundant. Function types should either be generic, and therefore uninstantiated (`ft_tparams` is non-empty), as is the case for generic function and generic method decls, or non-generic (`ft_tparams` is empty).

At present, `ft_tparams` does not get eliminated when the function type is instantiated, which results in weird hover types such as `identity<T>(int):int` that are a mashup of generic and instantiated types. This diff does not tackle that issue.

Reviewed By: CatherineGasnier

Differential Revision: D59955325

fbshipit-source-id: 0a7a48e44cf60bc7c286276667e276e9a7a879cc