Push reversals into projections
Summary:
A symmetric projections is one where we decompose a subtype and a supertype in the same way during constraint simplification e.g. `Prj_tuple of int` records the decomposition of a subtype constraint between two tuple types into a new subtype constraint between the types at the given index.

Certain 'symmetric' projections out of types are not truly symmetric. For example in the presence of variadic parameters we may be projecting a funcion type at type parameters appearing a different indices.

Because of these differences, we should be reversing these kinds of projections

Reviewed By: mheiber

Differential Revision: D59855078

fbshipit-source-id: 158ea82bc3af5b01bdd979ce0ed0f3fbf0b23b25