overload F14Table::prehash to take hashes
Summary: The current interface is `F14HashToken F14Table::prehash(K const&) const`. This is helpful for users to pipeline an overall lookup operation into separate `prehash` and then lookup-with-hash-token steps. But there is a further refinement: allow the user to split up prehashing into computing the hash value of the key from the hasher in one step and then postprocessing that hash value into a hash-token in a second step.

Reviewed By: bmaurer

Differential Revision: D60047675

fbshipit-source-id: 6620e7993d6b97f73b7629ad288e1831bc83650d