demonomorphize interrupt-handler in SemiFuture::within
Summary:
Demonomorphize the interrupt-handler function which is currently a lambda defined within `SemiFuture::within`.

Currently, a single handler function in concept ends up being many different identical handler functions, since it is templated over the type parameter of the `SemiFuture`. Extract the interrupt-handler function so that it is no longer templated.

Reviewed By: DenisYaroshevskiy

Differential Revision: D23467008

fbshipit-source-id: 048935a0af03e67aa28bcb5e61c84c15e772a25d