fix compile error in CHM test under gcc
Summary: Under gcc it is enforced that lambdas are not default-constructible - even empty lambdas - although it seems clang is permissive. The use of a lambda as a the type of a hash function, without providing the lambda to be copied into the CHM, fails to compile. CHM does not take instances of the hash function (it is an incomplete implementation in this sense) but the simplest workaround is to change the type of this hash function to a function-local struct.

Reviewed By: dmm-fb

Differential Revision: D60125349

fbshipit-source-id: a1f5d446562d457c024d1bfbac10932440dd4873