Delay error format resolution
Summary:
We were defaulting to Highlighted error formatter from option parsing stage. This has two problems:
1. We don't know when we defaulted vs the user explicitly set to `Highlighted`
2. There are multiple places that specify the default.

This diff refactors error format threading so that it is threaded as optional.

This diff is a nop. In the next diff we'll allow tools that default to `Plain` error format to override this behaviour by setting `--error-format` explicitly.

Reviewed By: CatherineGasnier

Differential Revision: D60182288

fbshipit-source-id: 1230c181e151a916981254625c755d1c8f105a20