Don't reestablish interactivity repeatedly
Summary:
To print errors and to show spinner we use the `--from` option's value to determine if twe're in an interactive context or not. Currently, this is either when `--from` is empty or when it is `"[sh]"`.

To prevent these going out of sync, I hoisted them to `clientArgs.ml` and pass around `is_interactive` boolean value instead.

This diff is a nop

Reviewed By: CatherineGasnier

Differential Revision: D60174393

fbshipit-source-id: e4e00d3d8c858374980e9b6d8d6baf3db5a63016