Make the py3 generator create empty files for normal fiels when in auto_migrate mode
Summary:
This is for the next diff in this stack, which changes the way that the auto_migrate build is implemented. Since it's a buck select now, the gencode needs to be changed to always spit out a given file name (or else buck complains that the file is missing) for the auto_migrate version of a given target, even if it isn't enabled yet. This shouldn't affect the actual build because these files aren't getting consumed by buck, the only cost is that the mustache template generator spits out a empty file that gets discarded later.

The primary change is that py3_mstch_generator no longer generates different sets of files based on whether or not auto_migrate is enabled, but now the templates themselves will be empty if they aren't supposed to be generated in the given migration mode.

Reviewed By: yoney

Differential Revision: D60195030

fbshipit-source-id: 95b58b27536ca89beea457ebd3b5793cde1c50a3