Invoke ClientInterceptors for sync_ AsyncClient functions
Summary:
This diff adds generated code for the `sync_*` variants of client functions to invoke `ClientInterceptorBase::{onRequest,onResponse}`.

Note that `sync_complete_*` is *not* covered — it's marked deprecated and only used by memcache.

Similar to `ServiceInterceptor`, the logic we want here is:
* Run all `onRequest` callbacks
* If no exceptions, make the RPC. Otherwise throw the exception.
* Run all `onResponse` callbacks. Otherwise throw the exception.
* If no exceptions, return the RPC result (could be a value or an exception)

Currently, this implementation is only for regular RPCs but more to follow in future diffs.

The structure of the `sync_*` implementation is very similar to the `co_*` implementation except that we have to block the calling thread.

Reviewed By: sazonovkirill

Differential Revision: D59835312

fbshipit-source-id: abe5050b44775e1b559e95881cdb25658513c63f