Factor out logic in unprefixed AsyncClient functions into a helper function
Summary:
Currently, `future_*` and `semifuture_*` implementations of `AsyncClient` delegate to the unprefixed (raw `RequestChannel`) variant. This variant is almost never used and not worth implementing `ClientInterceptor`s for (it would be callback hell).

This diff splits out the common part of the unprefixed function so it can be used to implement `future_*` and `semifuture_*` in a standalone way.

Thus, this diff has no observable side-effects.

Reviewed By: iahs

Differential Revision: D59835311

fbshipit-source-id: 58b690b0c52393d935830dc6c4faeaedb378c321