Refactor Future-based RequestChannel callbacks
Summary:
Context: the low level API for client-generated is callback based (`RequestCallback`). Higher level APIs such as `future_*` variants in generated `AsyncClient` objects are implemented by a `RequestCallback` which sets `folly::Promise` state.

This diff has two desired outcomes:
1. Alongside the result of the RPC, fulfill the promise with `ClientReceiveState`. This state contains `ContextStack` which manages the lifetimes of `ClientInterceptor` storage which we need to invoke post-RPC but before returning to the caller.
2. Use modern C++ features to DRY the code (e.g. remove the specialization for `folly::Unit`). The result is much more descriptive code.

In terms of performance, there is an additional `malloc` from the `FutureCore` allocated because of the continuation that calls `CallbackHelper::extractValue` in generated code. This is on par with using `Future`-based asynchrony. Furthermore, `future_*` variants are discouraged since we have `SemiFuture` now.

Reviewed By: AkramaMirza

Differential Revision: D59835310

fbshipit-source-id: fb2f85c42a4ed26aab63268a0f743fd2c46a5aa1