implement OOM multiplier to allow better chance of survival during OOM killing
Summary:
Some requests claim that they should preferrably not be killed (and consequently, other requests should be killed when the host is very low on memory).
This sounds unfair, but there are legitimate cases (e.g., unretriable critical requests).
So we implement an extension function set_oom_multiplier, which allows the request more memory budget to the OOM killer.

Requests setting a high multiplier should be charged more in cost attribution (to be implemented in a future diff).

Reviewed By: jtwarren

Differential Revision: D60366329

fbshipit-source-id: 34973cf56f0175b8c591991c2af2bb813652016c