upgrade to rocket server
Summary:
The upgrade to rocket server, accepts header protocol and then checks if the first message upgrades the connection to rocket and is ready to communicate in rocket or header, which ever the client prefers.

Integrating this in the rsocket library, took a bit of trickery
We need to swap out simply passing in a listener for implementing the transport.ServerTransport which acts as a listener factory.
Inside the `Listen(ctx context.Context, notifier chan<- bool)` method, we could call the listener ourselves, which gave us access to the net.Conn, which was hidden away from us inside of the rsocket library before.

We needed this net.Conn to communicate via HeaderProtocol with possible old clients. We copied the logic from the simple_server.go now header_server.go and adjust it to handle a single or multiple header requests. Now we could either:
* handle the rocket requests by falling back to the rsocket library.
* simply be a normal header server, yes we could now technically also delete header_server.go
* or act as an upgradeToRocket server, by reading the first request via headerProtocol and check if we were upgraded. If we were upgraded we fall back to the processing requests with the rsocket library's acceptor, that inside it contains the processorContext function, that we passed to the rsocket library's Server on startup OR if we weren't upgraded we continue to process header requests.

The upgrade procedure is checked inside the rocketUpgradeProcessor, which acts as a fall through to the user's processor. It simply checks for the "upgradeToRocket" method call, and if it receives it, it marks itself as upgraded, if not it falls through to the user's processor.

Reviewed By: echistyakov

Differential Revision: D60237423

fbshipit-source-id: 9204ac8b7e96a9c1fcb5759db9f2f1309195338f