Support unions in filter
Summary:
As title.

If an allMask is specified on a union-field, we simply copy the union into the result (semantically equivalent to copying the field)

If a partial mask is specified on a union field, we recurse into the child field, and only copy the result into the union if the recursion produces a non-empty result (maintain same semantics as optional field).

Reviewed By: thedavekwon

Differential Revision: D59550329

fbshipit-source-id: 09c28cac00bfbe27839830a3123d23fe21fe9b34