Expose user-returned value from ServiceHandler in ServiceInterceptor::onResponse
Summary:
`ServiceInterceptor::onResponse` now has access to the value returned from the user defined `ServiceHandler` (when available). We recently added the ability of introspect thrown exceptions. This extend the functionality to all outputs.

Like `onRequest`, we use `TypeErasedRef` to expose a type-safe way to access the returned value by the implementor of `ServiceInterceptor`.

See the added docs for more details.

Reviewed By: sazonovkirill

Differential Revision: D60469391

fbshipit-source-id: 3fe6c93abe51b96f4ffd8c0ea983de3c3856f9d7