Prevent premature short-circuiting of SFINAE in cursor-based serializer
Summary:
Compilers disagree on whether `enable_if_t<false && ...>` is SFINAEable. In particular, this arises in the cursor-based serializer:

```
  template <typename Ident>
  using enable_string_view = typename std::enable_if_t<
      Contiguous && type::is_a_v<type_tag<Ident>, type::string_c>,
      int>;
```
when `Contiguous` is `false`.

MSVC (which is usually wrong) gives the following error in this case:

```
CursorBasedSerializer.h(220): error C2938: 'std::enable_if_t<false,int>' : Failed to specialize alias template
```

Fix (or at least workaround) it by swapping the operands of `&&` so that it is not short circuited prematurely.

Reviewed By: avalonalex, thedavekwon

Differential Revision: D60478917

fbshipit-source-id: 4bd666b932d5c8065f133035a355131c9936f60f