Partially solve lower bounds on type variables when doing strip_dynamic
Summary:
Our implementation of the rule `C<~string> <: ~C<string>` and other aspects of sound-dynamic relies on spotting "~t" in types. Currently this is rather syntactic, and brittle, and will miss opportunities such as a lower bound `C<#0> <: ~C<string>` if `~string <: #0` is a constraint on the type variable.

In the code for `try_strip_dynamic` we can do a little early solving, in the special case of `dynamic, t1, ..., tn <: #0`, by solving `#0 := ~#1` for fresh type variable `#1` such that `t1, ..., tn <: #1`, thus exposing the "like-type" in the syntax of the type.

Differential Revision: D60447987

fbshipit-source-id: 755251e66bfc196c691342d071213c158b17b047