Add basic shape support
Summary:
This diff adds basic support for closed and open shapes with 0/1/2 fields. The field keys are chosen between "a" and "b".

Currently, this is pretty expensive especially while enumerating subtypes of shapes, so we simplify things by not adding extra fields when the shape is not open and not permuting fields.

Reviewed By: dlreeves

Differential Revision: D60352705

fbshipit-source-id: eab6345475f86baf450568140ba0c0d7cf9fb41a