Generalise tuple support to n-ary tuples
Summary:
See the title.

Sadly, this diff doesn't support nullary tuples because they are not denotable. We generate 1/2/3-tuples only.

In this diff, we also change case type disjointness machinery. Specifically, instead of weakening the type to a single type we weaken it to multiple possible types. This is to handle the fact that 1-tuples and 2-tuples are not disjoint and also do not share a _precise_ super type.

Reviewed By: dlreeves

Differential Revision: D60344455

fbshipit-source-id: bc2fd83e5f29fcae018c7e61739ce30c32ec6c0f