Add Refinement_counter
Summary:
Added a new map-reduce service that can be used to check progress on fully replacing the current refinement logic with the new constraint-based approach we have been working on. Most of the code is straightforward and I wrote emulating the other map-reduce services that are already implemented. Logically we look for `_ is T` expression nodes and check if `T` is a type we can produce a predicate for.

I changed the result of `Typing_refinement.TypePredicate.of_ty` to return `(type_predicate, string) Result.t` instead of `Option.t`. This allows to report a rudimentary reason for not being able to produce a predicate.

To enable the map-reduce you can set the log_level of `refinement_counter` equal to 1, 2 or 3:
1. Will compute a count of legacy vs. new refinements, along with counts that group legacy refinements based on reason why they cannot use the new refinement
2. Will do the same as 1, but also collect positions of all legacy refinements EXCEPT due to class (the predominate reason why we still use legacy refinement). Useful for debugging
3. Will do the same as 2, but also include positions for class based refinements

Reviewed By: viratyosin

Differential Revision: D60263237

fbshipit-source-id: 0ca034b4bf983124d60a679d091b9c3cfe2a89b0