let Range members take values by const&
Summary:
Various member functions, including `find` and `starts_with`, accept a `value_type` argument. But this interface is dangerous since it is possible for `value_type` to be non-copyable or to be expensively-copyable.

We change these members to take `value_type const&` but are careful not to let them take `const_reference` since the latter may not be the same as `value_type const&` or convertible from `value_type`.

Reviewed By: luciang, vitaut

Differential Revision: D60537917

fbshipit-source-id: f24dc9f41b3ef5ef4b3461e502612bb4e349af36