Co-op load-elim machinery for jmp elim
Summary:
Pretty simple and limited. Can work on improving but this fixes the current issue.

Other ideas:
1. Add `IntSpec` specialization to type-system
 Would probably work but on the edges, do you refine an SSATmp? (You gen an `AssertType`. But, how do you prevent this from being DCE'd before you can use it?

2. A new pass where we track SSA locations {ssatmp -> zero/nz}
Could work, but we already track SSATmps in a flow-sensititve way in load-elim. We should refactor load-elim to make it more generic. We also already take into account memory effects in load-elim, which wasn't used in this diff but is nice.

Reviewed By: ricklavoie

Differential Revision: D59887249

fbshipit-source-id: 7e3d6ea9cd4cb8f7406b0263b628457b0b75a335