Make exit when speculating and have no guess what the Class* is
Summary:
If are speculating, and we don't know a class, there are 2 options:
1.  (likely) The class has not been loaded yet. We should break the translation and load the class so we can start speculating immediately.
2. The class does not exist or has some error while loading. In this case we will fatal the request, which would've happened anyway.

If we are not speculating and don't know a class, that probably means we have a `Class*`, but we just can't guarantee its not going to change. In this case, if we did break, we'd generate infinite translations since we'd break every time we hit one of these bytecodes, load, and then re-jit the same thing, since loading the class is not going to help.

Reviewed By: ricklavoie

Differential Revision: D60138154

fbshipit-source-id: 9b51cdb13709efe20fd4248806e8104b36ea1542