Enable the use of buck select to toggle between auto-migrate and normal py3 mode instead of buck flags
Summary:
The previous prototype for enabling py3 auto-migrate required the use of buck flags, which had to be manually set in a cumbersome way via terminal

This changes the py3 bezel file to instead use buck select to toggle between an auto-migrated and default build tree in order to figure out what files are getting generated. This allows auto migrate to be toggled via PACKAGE files in a way that can be actually committed rather than set manually each run.

There's a demonstration of what this looks like for the end user in `thrift/lib/py3/test/auto_migrate_{on,off}`.

Credit to cxxxs for doing most of the work here in a previous prototype, I just fixed a few issues and cleaned it up for this diff.

Reviewed By: cxxxs

Differential Revision: D59123335

fbshipit-source-id: 591d2c34cab34d2e29a179bb5fbb8803d0b5b2a1