Add std::to_chars impl of folly::toAppend for floating points
Summary:
This diff uses `std::to_chars` for the implementation of `folly::toAppend `.  This is a step to removing the dependency on `libdouble-conversion`. libdouble-conversion is also used in string-to-float conversions. Another diff will replace that with std::from_chars

This implementation is enabled with a buck config switch, so clients can opt in. It's behind this switch because, unfortunately, the output of `std::to_chars` is not *exactly* identical to libdouble-conversion for all inputs and thus might unexpectedly break clients that depend on the formatting. NOTE: While the output from `std::to_chars` is not *exactly* identical, it is expected to represent the same IEEE-754 value as libdouble-conversion and thus not lose any precision.

Most of the code in this diffs adapts the output of `std::to_chars` to match the format of libdouble-convserion's output. It does this by parsing the decimal output and reformatting it in-place.

`_LIBCPP_AVAILABILITY_TO_CHARS_FLOATING_POINT` is used by libc++ to control the avalability of floating-point std::to_chars.

Reviewed By: yfeldblum

Differential Revision: D59700417

fbshipit-source-id: 133f14c3638029a375d6e27245c4510864721af6