another fix for Range operator== and follow-up tests
Summary:
{D60417647} changed `operator==(Range, Range)` when the `value_type` is integral or enumeration to use `memcmp` directly.

But it used the wrong size, which is fixed by {D60626149}. This diff adds tests that would have caught that problem.

And it did not limit to `Range` with only pointer iterators (could be extended to contiguous iterators once we have C++20). This diff limits the optimization to pointer iterators.

Differential Revision: D60630583

fbshipit-source-id: 67d6c06b488a8feb3a5c8cf666fc12b3d021aa0f