Remove path representation
Summary:
The path representation of a reason explain the subtype error in terms of directional flow. When using these flows to explain the problem, I often convert this back into an explanation containing:

1) An explanation of the subtyping derivation
2) An explanation of how the problematic type arrived at a term during typing

This diff is the first in a series which removes the 'path' representation and explanation in favor of this approach.

Reviewed By: CatherineGasnier

Differential Revision: D60583519

fbshipit-source-id: 875db0e4dd354399bdb4c892af7925151eca8c40