Add explicit `Solved` constructor
Summary:
We currently record the solution of a type variable the first time it is 'expanded' as a `Flow`. This gives us confusing provenance because the reason to which the solution applies is typically contained inside a `Flow`.

This diff introduces a `Solved` reason constructor and changes the two type variable witnesses to contain the `Tvid.t`. Doing so lets us match the reason in the solution to the correct reason no matter where it occurs

Reviewed By: madgen

Differential Revision: D60586346

fbshipit-source-id: f2375d51fa541ac77e2727db6dc1ad8e4515276f