Add derivation representation
Summary:
A reason is a record of a type provenance through:
1) typing
2) subtyping
3) solving

The reason representation can be thought about in terms of a failed subtyping derivation. Each 'derivation' may contain one or more sequences of subtyping 'steps' (each corresponding to some subtype constraint) with each step linked by some rule telling us how we got from the previous step. In the first step of the derivation the sub- and supertype must correspond to some expression or declaration. By unnesting the reason, we can recover this more intuitive representation as a derivation.

Reviewed By: madgen

Differential Revision: D60587133

fbshipit-source-id: bdaf1c3883ba0abf65d7a24de53726af7ddd83b0