Fix indexing type for FrozenRange
Summary:
The `ptrdiff_t` type allows for negative values, though negative indexing is not supported by this library.

Importantly, `ArrayLayout::View::indexPosition` expects a `size_t` for its second argument, meaning the the `index` argument here is silently being converted to an `unsigned` value. Other invocations of the `indexPosition` method in this class use the `index_` member variable, which is correctly defined as a `size_t`.

As a result, linters will discourage proper loop variable types when trying something along the lines of the following, because of the narrowing conversion:
```
for (size_t index = 0; index < myRange.size(); ++index) {
  const auto& element = myRange[index];
  // ...
}

Reviewed By: vitaut

Differential Revision: D60596946

fbshipit-source-id: 3b7979693fb57fb91b7bbc9d52fb3fc34a38cf1a