annotate generated classes
Summary:
Add annotations to generated classes, enums and typedefs so that Glean can index the mapping from generated to source entities (see linked task for  context).

This will be used to extend codehub navigation from c++ to thrift to types.

This is follows on what has been done for service functions D55445267.

Reviewed By: iahs

Differential Revision: D60142167

fbshipit-source-id: 5096c866f96531b5bd20288f902cf47c49f8e85e