fix indexing Thrift enums
Summary: D60450750  updated Hack indexer so it generates hack.HackToThrift facts, mapping generated enumerators to their source entity in a thrift file. It introduced a bug (which was apparent in the full run in the test plan, indexer randomly crashed),  which I attempted to fix in D60668291, unsuccessfully. This is the correct version.

Reviewed By: josefs

Differential Revision: D60723422

fbshipit-source-id: 5ad8f015865339c08e61ee13b074cffda58509e6