Move unnested rewrites to top of match
Summary:
As part of the provenance-aware subtyping work, we will be modifying simplification functions to only perform one step of a simplification.

This diff moves non-nested  'rewrite' rules to the top of the match. Most of these rewrites are for completeness or error reporting purposes but we have a couple which generate non-subtype constraints:
- dlreeves added a rule which rewrites subtype constraints with type predicates in supertype position into type-switch constraints
- there is a rule which rewrites a subtype constraint where we have a class with refinements in supertype position into the conjunction of the corresponding `has_type_member` constraints and a subtype constraint where the refinements are no longer contained in the class

Reviewed By: enetsee

Differential Revision: D56134605

fbshipit-source-id: 0aab3d04bf4434272eb0e69d0d7f841e392e75b3