Fix boundary conditions in accelerated hashing
Summary:
The hashing accelerators, as implemented, can over-read by one 64-bit register, which is usually benign. However, when this happens and the data being hashed ends on a page boundary and the next page is inaccessible (or unmapped), we get a SEGV.

Move boundary checking around a little so that we avoid this extra read when it's unnecessary.

Differential Revision: D60760907

fbshipit-source-id: 020e286eaf48477bb0f3c5dc4435ed28d47081ac