Handle Request priorities
Summary:
Doc on Thrift Server priorities :  https://www.internalfb.com/intern/staticdocs/thrift/docs/fb/server/prio-and-isolation/priorities-reference/

The end goal here is that important/critical thrift requests should have higher priority when queued on dispatcher.

Proxygen uses 2 level priorities - Important and Normal. The high priority endpoints are specified in config.hdf and if the request URI is listed in this set, it is assigned `Important` priority level on dispatcher. Users cannot override this by sending any special header.

Thrift Server uses 5 levels of priorities and these can be specified in IDL or using request headers.

Ideally, we should use the Thrift Server priority levels for HHVM Dispacther too. That requires fetching the static priorities specified in service IDL. Since metadata from IDL is not available within hhvm we have to resort to using config.hdf for improtant endpoints similar to Proxygen. With this we can duplicate the current behavior.

In addition, if user provides an override using header, it will override the config.hdf setting. Since header based priorities are very rarely used, it will be an overkill to use 5 levels. Let's use the same 2 levels of priority as Proxygen and in future if we have access to IDL metadata or change client behavior to always send the priority as header, we can change this to 5 levels.

Reviewed By: AkramaMirza

Differential Revision: D60603565

fbshipit-source-id: be4185b0bc913b51a6b7d2798ab5589db46fdb87